+++ lib/services/presence.go
@@ -38,6 +38,9 @@ type Presence interface {
 	// skipped to improve performance.
 	GetNodes(ctx context.Context, namespace string, opts ...MarshalOption) ([]Server, error)
 
+	// ListNodes returns a paginated list of registered servers.
+	ListNodes(ctx context.Context, namespace string, limit int, startKey string) (nodes []types.Server, nextKey string, err error)
+
 	// DeleteAllNodes deletes all nodes in a namespace.
 	DeleteAllNodes(ctx context.Context, namespace string) error